requirements.txt
2to3==1.0
agate==1.9.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
arrow==1.3.0
asgiref==3.9.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
binaryornot==0.4.4
bleach==6.2.0
bokeh==3.6.3
cachetools==5.5.2
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
colorcet==3.1.0
comm==0.2.2
contourpy==1.3.1
cookiecutter==2.6.0
cycler==0.12.1
daff==1.3.46
datashader==0.17.0
dbt-adapters==1.14.8
dbt-common==1.23.0
dbt-core==1.9.4
dbt-extractor==0.6.0
dbt-postgres==1.9.0
dbt-semantic-interfaces==0.7.4
debugpy==1.8.13
decorator==5.2.1
deepdiff==7.0.1
Django==5.2.4
djangorestframework==3.16.0
dotenv==0.9.9
executing==2.2.0
fastjsonschema==2.21.1
filelock==3.18.0
fonttools==4.56.0
fuzzywuzzy==0.18.0
h11==0.14.0
holoviews==1.20.2
humanize==4.12.2
idna==3.10
importlib-metadata==6.11.0
iniconfig==2.1.0
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
ipyvue==1.11.2
ipyvuetify==1.11.1
ipywidgets==8.1.6
isodate==0.6.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.14
kiwisolver==1.4.8
kmodes==0.12.2
leather==0.4.0
Levenshtein==0.27.1
linkify-it-py==2.0.3
llvmlite==0.44.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mashumaro==3.14
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
Mesa==2.1.2
Mesa-Viz-Tornado==0.1.3
more-itertools==10.7.0
msgpack==1.1.0
multipledispatch==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.0
ordered-set==4.1.0
packaging==24.2
pandas==2.2.3
panel==1.6.1
param==2.2.0
parsedatetime==2.6
parso==0.8.4
pathspec==0.12.1
pdb-tools==2.5.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.7
pluggy==1.5.0
prompt_toolkit==3.0.50
protobuf==5.29.4
psutil==7.0.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyct==0.5.0
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.19.1
pymdown-extensions==10.14.3
pynndescent==0.5.13
pyparsing==3.2.1
pytest==8.3.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-Levenshtein==0.27.1
python-slugify==8.0.4
pytimeparse==1.1.8
pytz==2025.1
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.3.0
RapidFuzz==3.12.2
reacton==1.9.1
referencing==0.36.2
requests==2.32.3
rich==14.0.0
rich-click==1.8.8
rpds-py==0.24.0
scikit-learn==1.6.1
scipy==1.15.2
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
snowplow-tracker==1.1.0
solara==1.45.0
solara-server==1.45.0
solara-ui==1.45.0
SQLAlchemy==2.0.40
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.46.2
text-unidecode==1.3
threadpoolctl==3.6.0
toolz==1.0.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
types-python-dateutil==2.9.0.20241206
typing==3.7.4.3
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
umap==0.1.1
umap-learn==0.5.7
urllib3==2.3.0
uvicorn==0.34.1
watchdog==6.0.0
watchfiles==1.0.5
wcwidth==0.2.13
webencodings==0.5.1
websockets==15.0.1
widgetsnbextension==4.0.14
xarray==2025.3.0
xgboost==3.0.0
xyzservices==2025.1.0
zipp==3.21.0
